requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.0